[wip] Handles (for client)

+++ src/compiler/backend/arch/x86_64/mod.rs
@@ -307,7 +307,7 @@ pub fn is_callee_saved(reg_id: MuID) -> bool {
 pub fn is_valid_x86_imm(op: &P<Value>) -> bool {
     use std::u32;
     match op.v {
-        Value_::Constant(Constant::Int(val)) if val <= u32::MAX as usize => {
+        Value_::Constant(Constant::Int(val)) if val <= u32::MAX as u64 => {
             true
         },
         _ => false
